Fix the FD_CLOEXEC bit
FD_CLOEXEC is not a bit position. It's a bit mask already.

See the definition in the kernel:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/include/uapi/asm-generic/fcntl.h

and the use of that constant in:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/fs/fcntl.c

The current code in sd-notify is a no-op. Bit 1 is not looked at by the kernel.
sd-notify does not touch bit 0.

Fix this by changing the mask.